+++ src/client/config/ado_net.rs
@@ -75,6 +75,12 @@ impl ConfigString for AdoNetConfig {
 mod tests {
     use super::*;
     use crate::client::AuthMethod;
+
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     use crate::EncryptionLevel;
 
     #[test]
@@ -363,6 +369,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_on() -> crate::Result<()> {
         let test_str = "encrypt=true";
         let ado: AdoNetConfig = test_str.parse()?;
@@ -373,6 +384,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_off() -> crate::Result<()> {
         let test_str = "encrypt=false";
         let ado: AdoNetConfig = test_str.parse()?;
@@ -383,6 +399,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_plaintext() -> crate::Result<()> {
         let test_str = "encrypt=DANGER_PLAINTEXT";
         let ado: AdoNetConfig = test_str.parse()?;
@@ -393,6 +414,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_missing() -> crate::Result<()> {
         let test_str = "";
         let ado: AdoNetConfig = test_str.parse()?;