+++ src/client/config/jdbc.rs
@@ -36,6 +36,12 @@ impl ConfigString for JdbcConfig {
 mod tests {
     use super::*;
     use crate::client::AuthMethod;
+
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     use crate::EncryptionLevel;
 
     #[test]
@@ -254,6 +260,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_on() -> crate::Result<()> {
         let test_str = "jdbc:sqlserver://my-server.com:4200;encrypt=true;";
         let jdbc: JdbcConfig = test_str.parse()?;
@@ -264,6 +275,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_off() -> crate::Result<()> {
         let test_str = "jdbc:sqlserver://my-server.com:4200;encrypt=false;";
         let jdbc: JdbcConfig = test_str.parse()?;
@@ -274,6 +290,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_plaintext() -> crate::Result<()> {
         let test_str = "jdbc:sqlserver://my-server.com:4200;encrypt=DANGER_PLAINTEXT;";
         let jdbc: JdbcConfig = test_str.parse()?;
@@ -284,6 +305,11 @@ mod tests {
     }
 
     #[test]
+    #[cfg(any(
+        feature = "rustls",
+        feature = "native-tls",
+        feature = "vendored-openssl"
+    ))]
     fn encryption_parsing_missing() -> crate::Result<()> {
         let test_str = "jdbc:sqlserver://my-server.com:4200;";
         let jdbc: JdbcConfig = test_str.parse()?;